+++ assets/javascript/destinations.js
@@ -1,0 +1,162 @@
+var destinations = {
+  beach: [
+    {
+      name: "Rio de Janeiro",
+      airportCode: "GIG",
+      country: "Brazil",
+      countryCode: "BR"
+    },
+    {
+      name: "Tel Aviv",
+      airportCode: "TLV",
+      country: "Israel",
+      countryCode: "IL"
+    },
+    {
+      name: "Paros",
+      airportCode: "PAS",
+      country: "Greece",
+      countryCode: "GR"
+    },
+    {
+      name: "Ko Samui",
+      airportCode: "USM",
+      country: "Thailand",
+      countryCode: "TH"
+    },
+    {
+      name: "Olbia",
+      airportCode: "OLB",
+      country: "Italy",
+      countryCode: "IT"
+    }
+  ],
+  hiking: [
+    {
+      name: "Okutama",
+      airportCode: "HND",
+      country: "Japan",
+      countryCode: "JP"
+    },
+    {
+      name: "Garmisch-Partenkirchen",
+      airportCode: "MUC",
+      country: "Germany",
+      countryCode: "DE"
+    },
+    {
+      name: "Grand Canyon",
+      airportCode: "GCN",
+      country: "United States",
+      countryCode: "US"
+    },
+    {
+      name: "Vancouver",
+      airportCode: "YVR",
+      country: "Canada",
+      countryCode: "CA"
+    },
+    {
+      name: "Mount Cook",
+      airportCode: "MON",
+      country: "New Zeland",
+      countryCode: "NZ"
+    }
+  ],
+  party: [
+    {
+      name: "Berlin",
+      airportCode: "TXL",
+      country: "Germany",
+      countryCode: "DE"
+    },
+    {
+      name: "Ibiza",
+      airportCode: "IBZ",
+      country: "Spain",
+      countryCode: "ES"
+    },
+    {
+      name: "Barcelona",
+      airportCode: "BCN",
+      country: "Spain",
+      countryCode: "ES"
+    },
+    {
+      name: "New York",
+      airportCode: "JFK",
+      country: "United States",
+      countryCode: "US"
+    },
+    {
+      name: "Mykonos",
+      airportCode: "JMK",
+      country: "Greece",
+      countryCode: "GR"
+    }
+  ],
+  sightseeing: [
+    {
+      name: "London",
+      airportCode: "LHR",
+      country: "England",
+      countryCode: "GB"
+    },
+    {
+      name: "Madrid",
+      airportCode: "MAD",
+      country: "Spain",
+      countryCode: "ES"
+    },
+    {
+      name: "Rome",
+      airportCode: "FCO",
+      country: "Italy",
+      countryCode: "IT"
+    },
+    {
+      name: "Hong Kong",
+      airportCode: "HKG",
+      country: "Hong Kong",
+      countryCode: "HK"
+    },
+    {
+      name: "Tokyo",
+      airportCode: "HND",
+      country: "Japan",
+      countryCode: "JP"
+    }
+  ],
+  ski: [
+    {
+      name: "Niseko",
+      airportCode: "CTS",
+      country: "Japan",
+      countryCode: "JP"
+    },
+    {
+      name: "Zermatt",
+      airportCode: "SIR",
+      country: "Switzerland",
+      countryCode: "CH"
+    },
+    {
+      name: "La Clusaz",
+      airportCode: "GVA",
+      country: "France",
+      countryCode: "FR"
+    },
+    {
+      name: "Sankt Anton am Arlberg",
+      airportCode: "INN",
+      country: "Austria",
+      countryCode: "AT"
+    },
+    {
+      name: "Revelstoke",
+      airportCode: "YLW",
+      country: "Canada",
+      countryCode: "CA"
+    }
+  ]
+};
