+++ index.html
@@ -53,9 +53,9 @@
       <div class="row column">
         <p class="lead1">What do you plan to do on your next trip?</p>
       </div>
-      <div class="row small-up-1 medium-up-2 large-up-3">
+      <main class="row small-up-1 medium-up-2 large-up-3" id="js-screenContainer">
         <div class="column">
-          <div class="callout" js-activity="beach">
+          <div class="callout activity" js-activity="beach">
             <p>
               <img
                 src="assets/images/beach.jpg"
@@ -70,7 +70,7 @@
           </div>
         </div>
         <div class="column">
-          <div class="callout" js-activity="hiking">
+          <div class="callout activity" js-activity="hiking">
             <p>
               <img
                 src="assets/images/hiking.jpg"
@@ -85,7 +85,7 @@
           </div>
         </div>
         <div class="column">
-          <div class="callout" js-activity="party">
+          <div class="callout activity" js-activity="party">
             <p>
               <img
                 src="assets/images/party.jpg"
@@ -100,7 +100,7 @@
           </div>
         </div>
         <div class="column">
-          <div class="callout" js-activity="sightseeing">
+          <div class="callout activity" js-activity="sightseeing">
             <p>
               <img
                 src="assets/images/sightseeing.jpg"
@@ -115,7 +115,7 @@
           </div>
         </div>
         <div class="column">
-          <div class="callout" js-activity="ski">
+          <div class="callout activity" js-activity="ski">
             <p>
               <img
                 src="assets/images/ski.jpg"
@@ -129,7 +129,7 @@
             </p>
           </div>
         </div>
-      </div>
+      </main>
       <footer>
         <div class="row">
           <div class="medium-12 columns">
@@ -145,6 +145,6 @@
       <script>
         $(document).foundation();
       </script>
-    <script type="text/javascript" src="assets/javascript/flights.js"></script>
-    <script type="text/javascript" src="assets/javascript/weather.js"></script>
+    <script type="text/javascript" src="assets/javascript/destinations.js"></script>
+    <script type="text/javascript" src="assets/javascript/main.js"></script>
 </html>
